% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFSurv_est.R
\name{FFSurv_est}
\alias{FFSurv_est}
\title{Flexible Feed-Forward Estimation in Discrete-time Panel Survival Models}
\usage{
FFSurv_est(
  X,
  intercept = T,
  psi = 1,
  init_params = NA,
  n_iter = 10000,
  bootstrap = F,
  bootstrap_n_iter = 1000,
  n_bootstrap = 1000
)
}
\arguments{
\item{X}{A data frame that must contain the following columns:
* 'id': a numeric or character column for unique identification of individuals.
* 'response': a numeric column representing the response variable.
* 'censor': a numeric binary column indicating whether the observation was right-censored or not (1 = censored, 0 = not censored).
Other columns can be included and will be treated as additional covariates.}

\item{intercept}{Logical. If TRUE, an intercept term is included in the model. Default is TRUE.}

\item{psi}{Numeric. Parameter for the grouping interval. Default is 1.}

\item{init_params}{Numeric vector. Initial values for the parameter estimates.
Not generally recommended to set manually as some initial values violate the
assumptions of the model and can result in bogus results. Default is NA.}

\item{n_iter}{Numeric. The maximum number of iterations for the likelihood optimization. Default is 1e4.}

\item{bootstrap}{Logical. If TRUE, bootstrap resampling is performed for statistical inference. Default is FALSE.}

\item{bootstrap_n_iter}{Numeric. The maximum number of iterations for the bootstrap process. Default is 1e3.}

\item{n_bootstrap}{Numeric. The number of bootstrap resamples to take. Default is 1e3.}
}
\value{
A list containing:
* Coefficients: a data frame containing the estimated coefficients and their associated standard errors, confidence intervals, and p-values.
* Frailty: a data frame showing the estimated shape and rate parameters for the frailty distribution.
* Baseline Hazard Plot: a plot object showing the estimated baseline hazard function.
}
\description{
This function implements the maximum likelihood method for discrete time panel
survival models that is proposed in the paper: "Bateni[2023]: Flexible Feed-Forward Estimation
in Discrete-time Panel Survival Models". The model allows for right-censoring, unobserved
heterogeneity and grouped responses. It estimates the effect of covariates in a panel setup
where individuals may undergo multiple duration spells.
}
\examples{
\dontrun{
  results <- FFSurv_est(X = data, intercept = TRUE, psi = 1, n_iter = 1e4)
}

}
\references{
Bateni[2023]: Flexible Feed-Forward Estimation in Discrete-time Panel Survival Models.
}
