% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFSurv_gen.R
\name{FFSurv_gen}
\alias{FFSurv_gen}
\title{Generate Survival Data for Flexible Feed-Forward Estimation}
\usage{
FFSurv_gen(
  n_person,
  n_incident,
  n_features = 3,
  beta = c(-1, 0.4, -1, 1),
  het = c("gamma", 1, 1),
  X_features = list(dist = c("normal", "normal", "dummy"), mean = c(5, 5, 0.5), sd =
    c(0.2, 0.2, NA)),
  baseline_hazard = list(formula = "y = log(1+x/125)", max = 100),
  psi = 1,
  censoring = list(mode = "fixed", val = 80, ratio = NA, sd = NA)
)
}
\arguments{
\item{n_person}{Integer. The number of individuals to generate data for.}

\item{n_incident}{Integer or vector. The number of incidents per person. If a single
integer is given, the same number of incidents is generated for each person.
If a vector is given, it should have length equal to n_person, specifying the number of
incidents for each person individually.}

\item{n_features}{Integer. The number of covariate features to include. Default is 3.}

\item{beta}{Numeric vector. The coefficients for the covariate features. Default is c(-1, .4, -1, 1).}

\item{het}{Vector. Specifies the distribution of unobserved heterogeneity.
The first element is a character string indicating the distribution ('gamma', 'lognorm', 'dual', 'uniform'),
and the remaining elements are parameters for the distribution.
  * 'gamma': Gamma distribution, with shape and scale parameters. Example: c('gamma', 1, 1)
  * 'lognorm': Log-normal distribution, with mean and standard deviation parameters. Example: c('lognorm', 1, 1)
  * 'dual': Binary distribution, with the probability of success and two numeric values
            representing the two possible outcomes. Example: c('dual', 0.5, 1, 5)
  * 'uniform': Uniform distribution, with minimum and maximum parameters. Example: c('uniform', 1, 5)
Default is c('gamma',1,1).}

\item{X_features}{List. Specifies the distributions of the covariate features.
The list has three elements:
  * 'dist': A character vector indicating the distribution of each feature ('normal' or 'dummy').
  * 'mean': A numeric vector indicating the mean of each feature.
  * 'sd': A numeric vector indicating the standard deviation of each feature. This can be NA for 'dummy' features.
Default is list( 'dist' = c('normal','normal','dummy'), 'mean' = c(5,5,0.5), 'sd' = c(.2,.2,NA) ).}

\item{baseline_hazard}{List. The function defining the baseline hazard and its maximum value.
The list has two elements:
  * 'formula': A formula according to which the baseline hazard function is constructed. It has to be in the form of "y = f(x)"
               with f(.) being the desired function.
  * 'max': The maximum value for the baseline hazard.
Default is list('formula' = 'y = log(1+x/125)', 'max' = 100).}

\item{psi}{Numeric. The grouping interval. Default is 1.}

\item{censoring}{List. Specifies the censoring mechanism. The first element 'mode' can be 'fixed', 'random', or 'None'.
  * 'fixed': Fixed censoring time. Specified by 'val' (fixed censoring time at a certain value) or 'ratio' (quantile for censoring time).
  * 'random': Random censoring times. Specified by 'val' (mean censoring time) and 'sd' (standard deviation).
  * 'None': No censoring.
Default is list( 'mode' = 'fixed', 'val' = 80, 'ratio'=NA, 'sd' = NA).}
}
\value{
A data frame with columns for 'id', 'response', 'censor', 'intercept',
and the covariate features.
}
\description{
This function generates survival data according to specified parameters.
The generated data can be used for flexible feed-forward estimation in discrete-time panel
survival models, including models with right-censoring, unobserved heterogeneity,
and grouped responses.
}
\examples{
\dontrun{
  data <- FFSurv_gen(n_person = 100, n_incident = 1, n_features = 3,
                     beta = c(-1, .4, -1, 1), het = c('gamma',1,1),
                     X_features = list( 'dist' = c('normal','normal','dummy'),
                                        'mean' = c(5,5,0.5), 'sd' = c(.2,.2,NA) ),
                     baseline_hazard = list('formula' = 'y = log(1+x/125)', 'max' = 100),
                     psi = 1,
                     censoring = list( 'mode' = 'fixed', 'val' = 80, 'ratio'=NA, 'sd' = NA))
}

}
\references{
Bateni[2023]: Flexible Feed-Forward Estimation in Discrete-time Panel Survival Models.
}
